MES-548 - Develop Algorithm for Sequential Image Processing

+++ src/main/java/com/alcegory/mescloud/azure/service/ContainerManagerServiceImpl.java
@@ -101,7 +101,15 @@ public class ContainerManagerServiceImpl implements ContainerManagerService {
     }
 
     private void checkImageOccurrencesToRemove(ContainerInfoDto containerInfoDto, ImageAnnotationDto uploadedImageAnnotationDto) {
+        if (containerInfoDto == null || containerInfoDto.getImageAnnotationDto() == null) {
+            return;
+        }
+
         String image = containerInfoDto.getImageAnnotationDto().getData().getImage();
+        if (image == null) {
+            return;
+        }
+
         int imageOccurrencesNotInitial = imageAnnotationService.countByImageAndStatusNotInitial(image);
         if (uploadedImageAnnotationDto != null && imageOccurrencesNotInitial >= 3) {
             publicContainerService.deleteBlob(image);
